backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.12.0
beanie==1.26.0
certifi==2025.11.12
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.115.5
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lazy-model==0.2.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
motor==3.6.0
boto3
opensearch-py==2.8.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.9.2
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
shellingham==1.5.4
SQLAlchemy==2.0.36
starlette==0.41.3
toml==0.10.2
tomli==2.3.0
typer==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
botocore==1.35.80
duckdb==1.1.3
pandas==2.2.3
slowapi
docker
pyarrow==18.1.0
s3fs==2024.12.0
openai>=1.0.0
